+++ frontend/src/app/App.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter } from "react-router-dom";
 
 import Roles from "./constants/roles";
 import RenderOnRole from "./components/routing/render-on-role";
-import { HwcrComplaintContainer } from "./components/complaints/hwcr/hwcr-complaint-container";
+import { HwcrComplaintContainer } from "./components/containers/complaints/hwcr/hwcr-complaint-container";
 import Layout from "./components/containers/layout";
 
 const App: FC = () => {
